Move param test to own file
